requirements.txt
django
django-tinymce
django-crispy-forms
django-recaptcha
pillow
six
psycopg2-binary
django-allauth
gunicorn
django-environ
django-safedelete
uuid
django-oauth-toolkit
django-filter
django-cors-headers
python-dotenv
djangorestframework
markdown 